1.02.00
Remake the game into a new version where the focus is cooking rather
than killing monster

There will be one character in the game and the objective of the game is
to gather ingredient around the cook according to the order in order to
server the order.

Fix a bunch of bug that cause the gemholder not able to replenished the
grid.

Have a working prototype of ingredient and order display.

Character are now able to gather ingredient round him to fulfill order.

Code for freshness of the ingredient. However, the freshness s not
properly display yet.

ToDoNext: Finish the order displaying.

Working on making the rat